moved App.js
